Renaming blockly_toolbox_flyout and blockly_trash_flyout as blockly_toolbox_ui and blockly_trash_ui, respectively. Moving BlockListUI, CategorySelectorUI, and DefaultVariableCallback to the .android.ui package.
