+++ blocklylib-core/src/main/java/com/google/blockly/android/control/BlocklyController.java
@@ -26,8 +26,8 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.view.ViewParent;
 
-import com.google.blockly.android.BlockListUI;
-import com.google.blockly.android.CategorySelectorUI;
+import com.google.blockly.android.ui.BlockListUI;
+import com.google.blockly.android.ui.CategorySelectorUI;
 import com.google.blockly.android.WorkspaceFragment;
 import com.google.blockly.android.clipboard.BlockClipDataHelper;
 import com.google.blockly.android.clipboard.SingleMimeTypeClipDataHelper;
